Use big integer for factorial output

+++ calc-worker-dotnet/CalcWorker/Api/JobDTO.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 
 namespace CalcWorker.Api
 {
@@ -7,7 +8,7 @@ namespace CalcWorker.Api
         public string Version { get; set; }
         public int Id { get; set; }
         public int? Input { get; set; }
-        public long? Output { get; set; }
+        public BigInteger? Output { get; set; }
         public DateTime? CalcStartedAt { get; set; }
         public DateTime? FinishedAt { get; set; }
         public string Status { get; set; }
